DESCW-607 picklist update optimization
