+++ backend/src/database/seeds/08_picker_options.js
@@ -1,3 +1,14 @@
+const {
+  yesNoOptions,
+  classification,
+  agreementType,
+  projectStatus,
+  projectType,
+  projectFunding,
+  projectRecoverable,
+  initiatedBy,
+} = require("../pick_options/dropdowns");
+
 exports.seed = function (knex) {
   const tables = {
     projects: "project",
@@ -11,315 +22,145 @@ exports.seed = function (knex) {
     contract_amendment: "contract_amendment",
   };
 
-  const yesNoOptions = [
-    {
-      value: "Yes",
-      label: "Yes",
-    },
-    {
-      value: "No",
-      label: "No",
-    },
-    {
-      value: "N/A",
-      label: "N/A",
-    },
-    {
-      value: null,
-      label: "",
-    },
-  ];
-
   const pickers = [
     {
       name: "classification",
       title: "Classification",
       description: "The classification type of the project.",
-      definition: {
-        dropDownValues: [
-          {
-            value: "Strategic",
-            label: "Strategic",
-          },
-          {
-            value: "Innovation",
-            label: "Innovation",
-          },
-          {
-            value: "Tactical",
-            label: "Tactical",
-          },
-          {
-            value: "Maintenance/Sustainment",
-            label: "Maintenance/Sustainment",
-          },
-          {
-            value: "Operational",
-            label: "Operational",
-          },
-          {
-            value: "Infrastructure",
-            label: "Infrastructure",
-          },
-          {
-            value: "Support for Strategic or Business Planning",
-            label: "Support for Strategic or Business Planning",
-          },
-          {
-            value: "Transformation",
-            label: "Transformation",
-          },
-        ],
-      },
+      definition: { dropDownValues: classification },
       associated_form: tables.projects,
     },
     {
       name: "agreement_type",
       title: "Agreement Type",
       description: "The agreement type of the project.",
-      definition: {
-        dropDownValues: [
-          {
-            value: "Project Charter",
-            label: "Project Charter",
-          },
-          {
-            value: "Other",
-            label: "Other",
-          },
-          {
-            value: "Partnership Agreement",
-            label: "Partnership Agreement",
-          },
-          {
-            value: "MOU",
-            label: "MOU",
-          },
-        ],
-      },
+      definition: { dropDownValues: agreementType },
       associated_form: tables.projects,
     },
     {
       name: "project_status",
       title: "Status",
       description: "The status of a project.",
-      definition: {
-        dropDownValues: [
-          {
-            value: "NewRequest",
-            label: "New Request",
-          },
-          {
-            value: "Active",
-            label: "Active",
-          },
-          {
-            value: "Cancelled",
-            label: "Cancelled",
-          },
-          {
-            value: "Complete",
-            label: "Complete",
-          },
-        ],
-      },
+      definition: { dropDownValues: projectStatus },
       associated_form: tables.projects,
     },
     {
       name: "ministry_id",
       title: "Client Ministry Name",
       description: "Client Ministry field",
-      definition: {
-        tableLookup: "ministry",
-      },
+      definition: { tableLookup: "ministry" },
       associated_form: tables.projects,
     },
     {
       name: "portfolio_id",
       title: "Portfolio Name",
       description: "Portfolio of the project.",
-      definition: {
-        tableLookup: "portfolio",
-      },
+      definition: { tableLookup: "portfolio" },
       associated_form: tables.projects,
     },
     {
       name: "fiscal_year",
       title: "Fiscal",
       description: "Fiscal Years",
-      definition: {
-        tableLookup: "fiscal_year",
-      },
+      definition: { tableLookup: "fiscal_year" },
       associated_form: tables.generic,
     },
     {
       name: "project_type",
       title: "Project Type",
       description: "The Project Type of a project.",
-      definition: {
-        dropDownValues: [
-          {
-            value: "External",
-            label: "External",
-          },
-          {
-            value: "Internal",
-            label: "Internal",
-          },
-        ],
-      },
+      definition: { dropDownValues: projectType },
       associated_form: tables.projects,
     },
     {
       name: "funding",
       title: "Funding",
       description: "The funding of a project.",
-      definition: {
-        dropDownValues: [
-          {
-            value: "Operational",
-            label: "Operational",
-          },
-          {
-            value: "Capital",
-            label: "Capital",
-          },
-          {
-            value: "Combination",
-            label: "Combination",
-          },
-        ],
-      },
+      definition: { dropDownValues: projectFunding },
       associated_form: tables.projects,
     },
     {
       name: "completed_by_contact_id",
       title: "Contact",
       description: "The Contact",
-      definition: {
-        tableLookup: "contact",
-      },
+      definition: { tableLookup: "contact" },
       associated_form: tables.projects,
     },
     {
       name: "recoverable",
       title: "Recovery Details",
       description: "The recoverable of a project.",
-      definition: {
-        dropDownValues: [
-          {
-            value: "Fully",
-            label: "Fully",
-          },
-          {
-            value: "Partially",
-            label: "Partially",
-          },
-          {
-            value: "Non-Recoverable",
-            label: "Non-Recoverable",
-          },
-        ],
-      },
+      definition: { dropDownValues: projectRecoverable },
       associated_form: tables.projects,
     },
     {
       name: "fiscal",
       title: "Fiscal",
       description: "Fiscal Years",
-      definition: {
-        tableLookup: "fiscal_year",
-      },
+      definition: { tableLookup: "fiscal_year" },
       associated_form: tables.generic,
     },
     {
       name: "hand_off_to_operations",
       title: "Hand Off to Operations",
       description: "Post implementation hand-off to operation completed.",
-      definition: {
-        dropDownValues: yesNoOptions,
-      },
+      definition: { dropDownValues: yesNoOptions },
       associated_form: tables.projects,
     },
     {
       name: "records_filed",
       title: "Records Filed",
       description: "Project documentation filled in accordance with records management.",
-      definition: {
-        dropDownValues: yesNoOptions,
-      },
+      definition: { dropDownValues: yesNoOptions },
       associated_form: tables.projects,
     },
     {
       name: "contract_ev_completed",
       title: "Contract Evaluation Completed",
       description: "Contract evaluation completed if applicable.",
-      definition: {
-        dropDownValues: yesNoOptions,
-      },
+      definition: { dropDownValues: yesNoOptions },
       associated_form: tables.projects,
     },
     {
       name: "contractor_security_terminated",
       title: "Contractor Security Terminated",
       description: "Contractor IDIR terminated / building passes returned.",
-      definition: {
-        dropDownValues: yesNoOptions,
-      },
+      definition: { dropDownValues: yesNoOptions },
       associated_form: tables.projects,
     },
     {
       name: "initiated_by",
       title: "Initiated By",
       description: "Who the change request was initiated by",
-      definition: {
-        dropDownValues: [
-          {
-            value: "GDX",
-            label: "GDX",
-          },
-          {
-            value: "Client",
-            label: "Client",
-          },
-        ],
-      },
+      definition: { dropDownValues: initiatedBy },
       associated_form: tables.change_request,
     },
     {
       name: "subcontractor_id",
       title: "Subcontractor",
       description: "Subcontractors",
-      definition: {
-        tableLookup: "subcontractor",
-      },
+      definition: { tableLookup: "subcontractor" },
       associated_form: tables.resource,
     },
     {
       name: "supplier_id",
       title: "Supplier",
       description: "Supplier",
-      definition: {
-        tableLookup: "supplier",
-      },
+      definition: { tableLookup: "supplier" },
       associated_form: tables.resource,
     },
     {
       name: "role_id",
       title: "Role",
       description: "The Role",
-      definition: {
-        tableLookup: "user_roles",
-      },
+      definition: { tableLookup: "user_roles" },
       associated_form: tables.users,
     },
     {
       name: "amendment_number",
       title: "Contract Type",
       description: "The Contract Type",
-      definition: {
-        tableLookup: "amendment_type",
-      },
+      definition: { tableLookup: "amendment_type" },
       associated_form: tables.contract_amendment,
     },
   ];
